refactor(assert): split identity equality helper
